+++ frontend/src/pages/Receipt/DomesticStock.tsx
@@ -16,7 +16,6 @@ import {
     formatNumber,
     createISODateKey
 } from '@/lib/utils/formatters';
-import { parseNumber } from '@/lib/utils/formatters';
 import { useReceiptData, useReceiptCalculations } from '@/hooks/receipt/useReceiptData';
 import {
     createYearOptions,
@@ -25,46 +24,7 @@ import {
     matchesYear,
     matchesAmounts
 } from '@/lib/utils/searchUtils';
-
-// CSVアイテムをDomesticStockDataに変換
-const parseCsvItem = (item: Record<string, unknown>): DomesticStockData => {
-    const account = String(item['口座'] || '');
-    const realizedPnL = parseNumber(item['実現損益[円]']);
-    // 特定口座の場合のみ税金を計算（利益がある場合のみ）
-    const isSpecificAccount = account.includes('特定');
-    const taxes = isSpecificAccount ? Math.floor(Math.max(0, realizedPnL) * TAX_RATE) : 0;
-    const realizedPnLAfterTax = realizedPnL - taxes;
-
-    const securityCode = String(item['銘柄コード']);
-    const securityName = String(item['銘柄名']);
-    // 銘柄コードをリンク形式で表示（/searchページに遷移）
-    const securityCodeLink = `<a href="/search?code=${securityCode}" class="security-code-link" style="color: #0d6efd; font-weight: 600;">${securityCode}</a>`;
-    // security_infoは後方互換性のため残す
-    const securityInfo = securityCodeLink;
-
-    return {
-        trade_date: new Date(item['約定日'] as string),
-        settlement_date: new Date(item['受渡日'] as string),
-        security_code: securityCode,
-        security_name: securityName,
-        security_info: securityInfo,
-        account,
-        shares: parseNumber(item['数量[株]']),
-        asked_price: parseNumber(item['売却/決済単価[円]']),
-        proceeds: parseNumber(item['売却/決済額[円]']),
-        purchase_price: parseNumber(item['平均取得価額[円]']),
-        realized_profit_and_loss: realizedPnL,
-        taxes,
-        realized_profit_and_loss_after_tax: realizedPnLAfterTax,
-    };
-};
-
-// 取引日でソート
-const sortByTradeDate = (data: DomesticStockData[]): DomesticStockData[] => {
-    return [...data].sort((a, b) =>
-        a.trade_date.getTime() - b.trade_date.getTime()
-    );
-};
+import { parseDomesticStockCsvItem, sortDomesticStockByTradeDate } from '@/features/receipt/parsers';
 
 // 日次データ集計
 const calculateDailyData = (domesticStockData: DomesticStockData[]): DomesticStockSummary[] => {
@@ -129,7 +89,7 @@ export const DomesticStock: React.FC<DomesticStockProps> = ({ csvData }) => {
     const [searchQuery, setSearchQuery] = useState('');
 
     // CSVデータを国内株式データ形式に変換
-    const domesticStockData = useReceiptData(csvData, parseCsvItem, sortByTradeDate);
+    const domesticStockData = useReceiptData(csvData, parseDomesticStockCsvItem, sortDomesticStockByTradeDate);
 
     // 日次データの集計
     const dailyData = calculateDailyData(domesticStockData);
